fix: 🐛 Move disco.scss into scss directory
`sass-loader` will concatinate all Sass partials into a single file when
it sees the `@import` directive. This broke fonts because the font urls
were placed in the `disco.scss` file. The paths referenced an
`../assets` directory, which is outside the scope of this project.
Moving `disco.scss` into the `scss` directory makes the urls work
properly for both the `scss-loader` and the generated CSS files.